+++ include/str.h
@@ -3,7 +3,7 @@
  * @brief uSched
  *        String helper interface header
  *
- * Date: 18-02-2015
+ * Date: 27-02-2015
  * 
  * Copyright 2014-2015 Pedro A. Hortas (pah@ucodev.org)
  *
@@ -31,7 +31,7 @@
 char *strrepl(const char *haystack, const char *needle, const char *rcontent);
 char *strreplall(const char *haystack, const char *needle, const char *rcontent);
 int strisnum(const char *s);
-int strrtrim(char *s, const char *trail);
+size_t strrtrim(char *s, const char *trail);
 
 #endif
