+++ src/usc/conn.c
@@ -3,7 +3,7 @@
  * @brief uSched
  *        Connections interface - Client
  *
- * Date: 25-02-2015
+ * Date: 27-02-2015
  * 
  * Copyright 2014-2015 Pedro A. Hortas (pah@ucodev.org)
  *
@@ -143,7 +143,7 @@ int conn_client_process(void) {
 		}
 
 		/* Send the first entry block */
-		if (panet_write(runc.fd, cur, usched_entry_hdr_size()) != usched_entry_hdr_size()) {
+		if (panet_write(runc.fd, cur, usched_entry_hdr_size()) != (ssize_t) usched_entry_hdr_size()) {
 			errsv = errno;
 			log_crit("conn_client_process(): panet_write() != %d: %s\n", usched_entry_hdr_size(), strerror(errno));
 			/* The payload size is in network byte order. We need to revert it before
@@ -166,7 +166,7 @@ int conn_client_process(void) {
 		cur->psize = ntohl(cur->psize) - (conn_is_remote(runc.fd) ? CRYPT_EXTRA_SIZE_CHACHA20POLY1305 : 0); /* Set the original payload size if the connection is remote. */
 
 		/* Read the session token into the session field for further processing */
-		if (panet_read(runc.fd, cur->session, sizeof(cur->session)) != sizeof(cur->session)) {
+		if (panet_read(runc.fd, cur->session, sizeof(cur->session)) != (ssize_t) sizeof(cur->session)) {
 			errsv = errno;
 			log_crit("conn_client_process(): panet_read() != sizeof(cur->session): %s\n", strerror(errno));
 			entry_destroy(cur);
@@ -211,7 +211,7 @@ int conn_client_process(void) {
 		memcpy(aaa_payload_data + sizeof(cur->session), cur->payload, cur->psize);
 
 		/* Send the authentication and authorization data along entry payload */
-		if (panet_write(runc.fd, aaa_payload_data, sizeof(cur->session) + cur->psize) != (sizeof(cur->session) + cur->psize)) {
+		if (panet_write(runc.fd, aaa_payload_data, sizeof(cur->session) + cur->psize) != (ssize_t) (sizeof(cur->session) + cur->psize)) {
 			errsv = errno;
 			log_crit("conn_client_process(): panet_write() != (sizeof(cur->session) + cur->psize): %s\n", strerror(errno));
 			entry_destroy(cur);
